plugins/CLIP/pyproject.toml
[project]
name = "CLIP"
version = "0.1.0"
description = "Semantic image search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "lancedb==0.22.0",
  "open-clip-torch>=2.32.0",
  "pyarrow==14.0.2",
  "rich>=14.3.3",
  "torch>=1.7.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
